Now using `preg_replace_callback()` instead of `preg_match_all()` with a foreach loop, as the `str_replace()` was causing multiple similar items to be escaped more than once
This fix was a **GREAT** idea conceived by @taliesinnz in GH issue #251.
She/he suggested I make the fix rather than he/she making a pull request
directly, so here it is. :)

This fix has the wonderful side-effect of increasing performance of
regular expression compilation, as it doesn't have to break apart the
string and replace each location multiple times.

Benchmarks show that this method is more linear (and consistent) in
processing time, so this is great. :+1: